:recycle: refactor : Query와 Command Service 역할 분리 완료
Relate to: #13
