+++ src/main/java/com/example/groutine/domain/member/service/MemberQueryService.java
@@ -25,37 +25,4 @@ public class MemberQueryService {
                 .orElseThrow(() -> new RestApiException(MemberErrorStatus.EMPTY_MEMBER));
     }
 
-    // 회원 저장
-    public Member saveEntity(Member member) {
-        return memberRepository.save(member);
-    }
-
-    public Member getCurrentMember() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        if (authentication == null || !authentication.isAuthenticated()) {
-            throw new RestApiException(MemberErrorStatus.UNAUTHORIZED); // 로그인 안함
-        }
-
-        Object principal = authentication.getPrincipal();
-
-        if (principal instanceof PrincipalDetails) {
-            return ((PrincipalDetails) principal).getMember();
-        }
-
-        throw new RestApiException(MemberErrorStatus.AUTHENTICATION_FAILED); // 로그인 정보를 확인할 수 없음
-
-    }
-
-    // 회원가입을 수행하는 함수
-    public MemberIdResponse signUp(Member member, MemberInfoRequest request) {
-        //이미 소셜 로그인 후, 인증 완료되면 멤버 엔티티는 생겨 있는 상태
-        //그 후 추가 정보를 입력받아 저장하는 메서드
-        Member loginMember = findById(member.getId());
-
-        // 기본 정보 저장 로직 작성 필요
-        loginMember.setName(request.getName());
-
-        return new MemberIdResponse(saveEntity(loginMember).getId());
-    }
 }
